changed dino game
