+++ day006/dino-game/src/App.jsx
