+++ day006/dino-game/src/components/modal.css
@@ -1,0 +1,22 @@
+dialog{
+    width: 50%;
+    position: fixed;
+    top: 60px;
+    z-index: 10;
+    border: none;
+    outline: none;
+    border-radius: 10px;
+    background-color: cadetblue;
+    color: white;
+}
+
+dialog button{
+    background: none;
+    border: none;
+    cursor: pointer;
+}
+dialog button::before{
+    content: "❌";
+    font-size: 1.3rem;
+
+}
