+++ day006/dino-game/src/main.jsx
