+++ day006/dino-game/vite.config.js
@@ -1,0 +1,6 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react"
+
+export default defineConfig({
+    plugins:[react()]
+})
